extract checkcode fetching method to a dedicated in function service
